Revert moving SpotifyAPIService.ts to brainzplayer
