+++ listenbrainz/webserver/static/js/src/playlists/Playlist.tsx
@@ -29,7 +29,7 @@ import {
 } from "../AlertNotificationsHOC";
 import APIServiceClass from "../APIService";
 import GlobalAppContext, { GlobalAppContextT } from "../GlobalAppContext";
-import SpotifyAPIService from "../brainzplayer/SpotifyAPIService";
+import SpotifyAPIService from "../SpotifyAPIService";
 import BrainzPlayer from "../brainzplayer/BrainzPlayer";
 import Card from "../components/Card";
 import Loader from "../components/Loader";
